Make the Aspell dictionary path configure for the Spellchecker library
The binary I compiled on Heroku by default is /tmp/aspell/lib/aspell-0.60. Vulcan compiles in the /tmp directory and I assume this is why the default is in the root of /tmp. In production on Heroku this should be /app/lib/aspell-0.60/. See "Packaging Binary Buildpack Dependencies" for the procedure I followed to build the binary:

https://devcenter.heroku.com/articles/buildpack-binaries